Add iOS test examples
