+++ ios/conf/ios-hybrid-properties.cfg
@@ -6,6 +6,6 @@ appium_app_strings: true
 automationName: Appium
 platformName: iOS
 deviceName: iPhone 6
-platformVersion: 9.1
+platformVersion: 8.3
 browserName:
-app: https://github.com/appium/javascript-workshop/blob/master/apps/HelloGappium.app.zip?raw=true
+app: https://github.com/appium/javascript-workshop/blob/master/apps/HelloGappium.app.zip?raw=true&fake=.zip
