+++ ios/conf/logging.conf
@@ -1,0 +1,46 @@
+[loggers]
+keys=root,toolium,remote_connection,pil
+
+[handlers]
+keys=consoleHandler,fileHandler
+
+[formatters]
+keys=simpleFormatter
+
+[logger_root]
+level=DEBUG
+handlers=consoleHandler,fileHandler
+
+[logger_toolium]
+level=DEBUG
+handlers=consoleHandler,fileHandler
+qualname=toolium
+propagate=0
+
+[logger_remote_connection]
+level=INFO
+handlers=fileHandler
+qualname=selenium.webdriver.remote.remote_connection
+propagate=0
+
+[logger_pil]
+level=INFO
+handlers=fileHandler
+qualname=PIL
+propagate=0
+
+[handler_consoleHandler]
+class=StreamHandler
+level=DEBUG
+formatter=simpleFormatter
+args=(sys.stdout,)
+
+[handler_fileHandler]
+class=handlers.RotatingFileHandler
+level=DEBUG
+formatter=simpleFormatter
+args=('%(logfilename)s','a',2000000,10)
+
+[formatter_simpleFormatter]
+format=%(levelname)-7s %(asctime)s [%(module)s] %(message)s
+datefmt=
