+++ ios/conf/properties.cfg
@@ -1,0 +1,43 @@
+[Driver]
+# Valid driver types: firefox, chrome, iexplore, edge, safari, opera, phantomjs, ios, android
+type: firefox
+# Configure local driver paths
+chrome_driver_path:
+explorer_driver_path:
+edge_driver_path:
+opera_driver_path:
+phantomjs_driver_path:
+# Driver options
+implicitly_wait: 5
+reuse_driver: false
+appium_app_strings: false
+
+[Server]
+enabled: false
+host:
+port:
+video_enabled: false
+
+[Capabilities]
+# Selenium capabilities: https://github.com/SeleniumHQ/selenium/wiki/DesiredCapabilities
+
+[AppiumCapabilities]
+# Appium capabilities: http://appium.io/slate/en/master/?ruby#appium-server-capabilities
+
+[VisualTests]
+enabled: true
+fail: false
+save: false
+complete_report: true
+baseline_name: {Driver_type}
+engine: pil
+
+[Jira]
+enabled: false
+execution_url:
+onlyifchanges: true
+summary_prefix:
+fixversion:
+labels:
+comments:
+build:
